requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
eval_type_backport==0.2.2
fastapi==0.135.1
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.11
importlib_metadata==8.7.1
jsonpath-python==1.1.4
lxml==6.0.2
mistralai==1.2.5
mypy_extensions==1.1.0
opentelemetry-api==1.39.1
opentelemetry-semantic-conventions==0.60b1
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.27.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.22
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.52.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
zipp==3.23.0
